+++ src/Entity/DailyOccurrence.php
@@ -24,7 +24,7 @@ class DailyOccurrence
     #[ORM\Column]
     private ?\DateTimeImmutable $end = null;
 
-    #[ORM\Column(length: 255)]
+    #[ORM\Column(length: 255, nullable: true)]
     private ?string $ticketPriceRange = null;
 
     #[ORM\Column(length: 255, nullable: true)]