+++ src/MessageHandler/GeocoderHandler.php
@@ -3,9 +3,9 @@
 namespace App\MessageHandler;
 
 use App\Exception\GeocoderException;
+use App\Message\DailyOccurrenceMessage;
 use App\Message\GeocoderMessage;
 use App\Message\IndexMessage;
-use App\Message\OccurrenceSplitterMessage;
 use App\Repository\AddressRepository;
 use App\Repository\EventRepository;
 use App\Service\Geocoder;
@@ -50,7 +50,7 @@ final class GeocoderHandler
         }
 
         // Send the event into the search index and at the same time to the occurrence splitter.
-        $this->messageBus->dispatch(new OccurrenceSplitterMessage($message->getEventId()));
-        $this->messageBus->dispatch(new IndexMessage($message->getEventId()));
+        $this->messageBus->dispatch(new DailyOccurrenceMessage($message->getEventId()));
+        //        $this->messageBus->dispatch(new IndexMessage($message->getEventId()));
     }
 }