+++ src/Model/DateTimeInterval.php
@@ -5,8 +5,8 @@ namespace App\Model;
 final class DateTimeInterval
 {
     public function __construct(
-        public ?\DateTimeImmutable $start = null,
-        public ?\DateTimeImmutable $end = null,
+        public \DateTimeImmutable $start,
+        public \DateTimeImmutable $end,
     ) {
     }
 }